POC crate refactor fast
keeps existing bench speeds...
