+++ benches/src/lib.rs
@@ -1,0 +1,1 @@
+// This file is intentionally empty to satisfy the crate requirements.
