+++ circuit/src/circuit.rs
@@ -1,9 +1,9 @@
 //! Wormhole Circuit.
 //!
 //! This module defines the zero-knowledge circuit for the Wormhole protocol.
-use amounts::{Amounts, AmountsTargets};
-use exit_account::{ExitAccount, ExitAccountTargets};
-use nullifier::{Nullifier, NullifierTargets};
+use crate::amounts::{Amounts, AmountsTargets};
+use crate::exit_account::{ExitAccount, ExitAccountTargets};
+use crate::nullifier::{Nullifier, NullifierTargets};
 use plonky2::{
     field::{goldilocks_field::GoldilocksField, types::Field},
     iop::witness::PartialWitness,
@@ -13,15 +13,8 @@ use plonky2::{
         config::PoseidonGoldilocksConfig,
     },
 };
-use storage_proof::{StorageProof, StorageProofTargets};
-use unspendable_account::{UnspendableAccount, UnspendableAccountTargets};
-
-pub mod amounts;
-pub mod exit_account;
-mod gadgets;
-pub mod nullifier;
-pub mod storage_proof;
-pub mod unspendable_account;
+use crate::storage_proof::{StorageProof, StorageProofTargets};
+use crate::unspendable_account::{UnspendableAccount, UnspendableAccountTargets};
 
 // Plonky2 setup parameters.
 pub const D: usize = 2; // D=2 provides 100-bits of security
@@ -118,7 +111,7 @@ impl WormholeCircuit {
     }
 }
 
-#[cfg(test)]
+#[cfg(any(test, feature = "testing"))]
 pub mod tests {
     use plonky2::plonk::proof::ProofWithPublicInputs;
