+++ circuit/src/gadgets.rs
