+++ circuit/src/lib.rs
@@ -1,0 +1,8 @@
+pub mod circuit;
+pub mod amounts;
+pub mod exit_account;
+pub mod nullifier;
+pub mod storage_proof;
+pub mod unspendable_account;
+pub mod inputs;
+pub mod gadgets;
