+++ circuit/src/nullifier.rs
@@ -1,15 +1,20 @@
 use plonky2::{
-    hash::{hash_types::HashOutTarget, poseidon::PoseidonHash},
+    hash::{
+        hash_types::HashOutTarget,
+        poseidon::PoseidonHash,
+    },
     iop::{
         target::Target,
         witness::{PartialWitness, WitnessWrite},
     },
-    plonk::{circuit_builder::CircuitBuilder, config::Hasher},
+    plonk::{
+        circuit_builder::CircuitBuilder,
+        config::Hasher,
+    },
 };
 
-use crate::prover::CircuitInputs;
-
-use super::{CircuitFragment, D, Digest, F, slice_to_field_elements};
+use crate::circuit::{CircuitFragment, D, Digest, F, slice_to_field_elements};
+use crate::inputs::CircuitInputs;
 
 // FIXME: Adjust as needed.
 pub const PREIMAGE_NUM_TARGETS: usize = 5;
@@ -112,12 +117,12 @@ pub mod test_helpers {
 #[cfg(test)]
 pub mod tests {
     use plonky2::{field::types::Field, plonk::proof::ProofWithPublicInputs};
-
+    
     use crate::circuit::{
         C,
-        nullifier::test_helpers::PREIMAGE,
         tests::{build_and_prove_test, setup_test_builder_and_witness},
     };
+    use crate::nullifier::test_helpers::PREIMAGE;
 
     use super::*;
