+++ prover/src/prover.rs
@@ -1,81 +1,20 @@
-//! Prover logic for the Wormhole circuit.
-//!
-//! This module provides the [`WormholeProver`] type, which allows committing inputs to the circuit
-//! and generating a zero-knowledge proof using those inputs.
-//!
-//! The typical usage flow involves:
-//! 1. Initializing the prover (e.g., via [`WormholeProver::default`] or [`WormholeProver::new`]).
-//! 2. Creating user inputs with [`CircuitInputs`].
-//! 3. Committing user inputs using [`WormholeProver::commit`].
-//! 4. Generating a proof using [`WormholeProver::prove`].
-//!
-//! # Example
-//!
-//! ```
-//! use wormhole_circuit::prover::{WormholeProver, CircuitInputs};
-//!
-//! # fn main() -> anyhow::Result<()> {
-//! # let inputs = CircuitInputs::default();
-//! let prover = WormholeProver::new();
-//! let proof = prover.commit(&inputs)?.prove()?;
-//! # Ok(())
-//! # }
-//! ```
 use anyhow::bail;
 use plonky2::{
     iop::witness::PartialWitness,
     plonk::{circuit_data::ProverCircuitData, proof::ProofWithPublicInputs},
 };
 
-use crate::circuit::{
-    C, CircuitFragment, CircuitTargets, D, F, WormholeCircuit,
+use wormhole_circuit::circuit::{C, D, F, WormholeCircuit};
+use wormhole_circuit::{
+    circuit::{CircuitFragment, CircuitTargets},
     amounts::Amounts,
     exit_account::ExitAccount,
     nullifier::{Nullifier, NullifierInputs},
     storage_proof::{StorageProof, StorageProofInputs},
     unspendable_account::{UnspendableAccount, UnspendableAccountInputs},
+    inputs::CircuitInputs,
 };
 
-/// Inputs required to commit to the wormhole circuit.
-#[derive(Debug)]
-pub struct CircuitInputs {
-    /// The amount sent in the transaction.
-    pub funding_tx_amount: u64,
-    /// Amount to be withdrawn.
-    pub exit_amount: u64,
-    /// The fee for the transaction.
-    pub fee_amount: u64,
-    /// Raw bytes of the nullifier preimage, used to prevent double spends.
-    pub nullifier_preimage: Vec<u8>,
-    /// Raw bytes of the unspendable account preimage.
-    pub unspendable_account_preimage: Vec<u8>,
-    /// A sequence of key-value nodes representing the storage proof.
-    ///
-    /// Each element is a tuple where the items are the left and right splits of a proof node split
-    /// in half at the expected childs hash index.
-    pub storage_proof: Vec<(Vec<u8>, Vec<u8>)>,
-    /// The root hash of the storage trie.
-    pub root_hash: [u8; 32],
-    /// The address of the account to pay out to.
-    pub exit_account: [u8; 32],
-}
-
-/// A prover for the wormhole circuit.
-///
-/// # Example
-///
-/// Setup prover, commit inputs, and then generate the proof:
-///
-/// ```
-/// use wormhole_circuit::prover::{WormholeProver, CircuitInputs};
-///
-/// # fn main() -> anyhow::Result<()> {
-/// # let inputs = CircuitInputs::default();
-/// let prover = WormholeProver::new();
-/// let proof = prover.commit(&inputs)?.prove()?;
-/// # Ok(())
-/// # }
-/// ```
 #[derive(Debug)]
 pub struct WormholeProver {
     circuit_data: ProverCircuitData<F, C, D>,
@@ -160,41 +99,13 @@ impl WormholeProver {
     }
 }
 
-#[cfg(any(test, feature = "testing"))]
-mod test_helpers {
-    use crate::circuit::{
-        nullifier,
-        storage_proof::test_helpers::{ROOT_HASH, default_proof},
-        unspendable_account,
-    };
-
-    use super::CircuitInputs;
-
-    impl Default for CircuitInputs {
-        fn default() -> Self {
-            let nullifier_preimage = hex::decode(nullifier::test_helpers::PREIMAGE).unwrap();
-            let unspendable_account_preimage =
-                hex::decode(unspendable_account::test_helpers::PREIMAGES[0]).unwrap();
-            let root_hash: [u8; 32] = hex::decode(ROOT_HASH).unwrap().try_into().unwrap();
-            Self {
-                funding_tx_amount: 0,
-                exit_amount: 0,
-                fee_amount: 0,
-                nullifier_preimage,
-                unspendable_account_preimage,
-                storage_proof: default_proof(),
-                root_hash,
-                exit_account: [0u8; 32],
-            }
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use super::{CircuitInputs, WormholeProver};
-
+    use wormhole_circuit::inputs::CircuitInputs;
+    use super::WormholeProver;
+    
     #[test]
+    #[cfg(feature = "testing")]
     fn commit_and_prove() {
         let prover = WormholeProver::new();
         let inputs = CircuitInputs::default();
