+++ verifier/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "wormhole-verifier"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+description.workspace = true
+license.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+plonky2.workspace = true
+poseidon-resonance.workspace = true
+wormhole-circuit = { path = "../circuit" }
+
+[dev-dependencies]
+wormhole-prover = { path = "../prover" }
+
+[features]
+testing = ["wormhole-circuit/testing", "wormhole-prover/testing"]
